bugs fix
